Update REAMDE and initial commit to ast.ml
The README more accurately describes the purpose of the project. More
work should be done on the brief, and the
installation/dependencies/usage sections are done yet started. A very
simple implementation of the ast is now usable. Could be updated to
include matches in the expression type. Other constructions
(equations/statements) should be in their own types.